Added situation variable and release
This variable lets us get into a situation that when we get "symbol" + "-" all symbol buttons disabled until an action rather than symbols happen.

With this update im satisfied. I checked all possible bugs and troubles with the calculator i made.  This is a released version from now on. I can call it v1.

If anyone read this and could find a bug just report or send a pull request thanks!